src/requirements.txt
boto3        # For AWS SDK
flask        # For local API testing (not needed for Lambda deployment)
psycopg2-binary